moved all static methods from ebeaninternal/AnnotationBase to io.ebean.util.AnnotationUtil
