+++ src/main/java/io/ebeaninternal/server/core/bootup/BootupClasses.java
@@ -16,7 +16,7 @@ import io.ebean.event.changelog.ChangeLogPrepare;
 import io.ebean.event.changelog.ChangeLogRegister;
 import io.ebean.event.readaudit.ReadAuditLogger;
 import io.ebean.event.readaudit.ReadAuditPrepare;
-import io.ebeaninternal.server.deploy.parse.AnnotationBase;
+import io.ebean.util.AnnotationUtil;
 import io.ebeaninternal.server.type.ScalarType;
 import org.avaje.classpath.scanner.ClassFilter;
 import org.slf4j.Logger;
@@ -475,6 +475,6 @@ public class BootupClasses implements ClassFilter {
    * Returns true if this class has the annotation (or meta annotation). Does not search recursively.
    */
   private boolean has(Class<?> cls, Class<? extends Annotation> ann) {
-    return AnnotationBase.findAnnotation(cls, ann) != null;
+    return AnnotationUtil.findAnnotation(cls, ann) != null;
   }
 }
