+++ src/main/java/io/ebeaninternal/server/deploy/parse/AnnotationSql.java
@@ -1,6 +1,7 @@
 package io.ebeaninternal.server.deploy.parse;
 
 import io.ebean.annotation.Sql;
+import io.ebean.util.AnnotationUtil;
 import io.ebeaninternal.server.deploy.BeanDescriptor;
 
 /**
@@ -15,7 +16,7 @@ public class AnnotationSql extends AnnotationParser {
   @Override
   public void parse() {
     Class<?> cls = descriptor.getBeanType();
-    Sql sql = AnnotationBase.findAnnotationRecursive(cls, Sql.class);
+    Sql sql = AnnotationUtil.findAnnotationRecursive(cls, Sql.class);
     if (sql != null) {
       descriptor.setEntityType(BeanDescriptor.EntityType.SQL);
     }
